requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
Django==4.2.5
django-channels==0.7.0
idna==3.4
msgpack==1.0.6
oauthlib==3.2.2
psycopg2-binary==2.9.9
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.5
gunicorn
psycopg2